% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fix_encoding}
\alias{fix_encoding}
\title{Conserta problemas de encoding}
\usage{
fix_encoding(x)
}
\arguments{
\item{x}{um vetor de caracteres}
}
\value{
o vetor x com os problemas corrigidos
}
\description{
não sei direito qual é o motivo, mas todos os nomes extraidos têm problemas
de encoding no meu computador, por isso existe essa função.
}

