% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fix_encoding_df}
\alias{fix_encoding_df}
\title{Conserta os problemas de encoding em data.frames}
\usage{
fix_encoding_df(x)
}
\arguments{
\item{x}{data.frame com problemas de encoding}
}
\value{
o data.frame x com os problemas corrigidos
}
\description{
não sei direito qual é o motivo, mas todos os nomes extraidos têm problemas
de encoding no meu computador, por isso existe essa função.
}

