% Generated by roxygen2 (4.0.1): do not edit by hand
\name{jup_disciplinas}
\alias{jup_disciplinas}
\title{disciplinas de uma unidade de ensino}
\usage{
jup_disciplinas(codigo)
}
\arguments{
\item{codigo}{da unidade de ensino da qual deseja-se obter a lista de disciplinas.}
}
\description{
dado o código de uma unidade de ensino, retorna a lista de disciplinas, com sigla, nome, datas de ativação e desativação
}

